Combined debug/release installation for Windows
This commit makes it possible to install both the debug
and release versions of the library into the same directory
on windows. This is accomplished by using different library
names for the two build types:

  * QBrowseButton[45]d - Qt4/5 Debug build
  * QBrowseButton[45]  - Qt4/5 Release build

The 'd' suffix is appended to all libraries and executables so
it's also possible to have two different versions of the plugin
for Qt Designer and the example program.

Signed-off-by: Peter Hille (png!das-system) <peter@das-system-networks.de>